requirements.txt
openai-whisper
gtts
sounddevice
scipy
python-dotenv
requests
